fix(health-sync): decode sync_state.retry_count as i32, un-panicking the PG cursor read SyncCursorRow carried retry_count as i64 but migrations_pg declares INTEGER (INT4); sqlx's strict PG decode rejects the width mismatch and Row::get panics inside get_sync_cursor. Unreachable until c054dc0a fixed the cursor *write* path (TEXT vs TIMESTAMPTZ binds) — the first row ever written then made every chat-pipeline refresh panic for that user (dev outage 2026-07-05, correlation ids 7df0a1b1 / 69807284; the dispatch panic boundary degraded it to the localized generic reply). Type the field i32 — sqlx decodes/binds native INT4 on PG and INTEGER on SQLite from the same field type (the c054dc0a pattern) — and swap the health_sync conversions to cast_unsigned/cast_signed. New test_parity_sync_cursor_roundtrip pins the full write-then-read cycle on both backends (red on pre-fix code with the exact production ColumnDecode, green after); create_test_server_resources hard-codes sqlite::memory: even under --features postgresql, so the parity file is the only test seam that actually exercises PG decode.

+++ crates/pierre-database/src/repositories/health.rs
@@ -39,8 +39,11 @@ pub struct SyncCursorRow {
     pub records_synced: i64,
     /// Error message from the last failed sync
     pub error_message: Option<String>,
-    /// Number of consecutive retry attempts
-    pub retry_count: i64,
+    /// Number of consecutive retry attempts. `i32` so the PG backend decodes
+    /// the INTEGER (INT4) column natively — sqlx's strict PG decode rejects an
+    /// `i64` read of INT4 (and `Row::get` panics on it); `SQLite` INTEGER
+    /// decodes either width.
+    pub retry_count: i32,
     /// When to attempt the next retry
     pub next_retry_at: Option<DateTime<Utc>>,
 }
